arch/m65832: Fix page table types for 64-bit PTEs
The M65832 MMU uses 64-bit page table entries:
- Bit 63: NX (No Execute)
- Bits 62:12: Physical Page Number (52 bits)
- Bits 11:0: Flags

Update page table definitions:
- pte_t, pgd_t, pgprot_t now use 'unsigned long long'
- PGD_ORDER and PTE_ORDER set to 1 (8KB = 2 pages per table)
- _PAGE_* flags use ULL suffix for 64-bit values
- _PFN_MASK properly masks 52-bit physical address
- Page allocations use correct order for 8KB tables

This matches the M65832 Linux Porting Guide specification:
- L1 table: 1024 entries × 8 bytes = 8KB
- L2 table: 1024 entries × 8 bytes = 8KB
- Total: 10+10+12 bit VA split (PGD index, PTE index, offset)

+++ arch/m65832/include/asm/page.h
@@ -28,18 +28,25 @@
 #define copy_user_page(to, from, vaddr, pg)	copy_page(to, from)
 
 /*
- * Page table entry types for type checking
+ * Page table entry types for type checking.
+ *
+ * M65832 uses 64-bit page table entries:
+ * - Bit 63: NX (No Execute)
+ * - Bits 62:12: Physical Page Number (52 bits)
+ * - Bits 11:0: Flags (Present, Writable, User, etc.)
+ *
+ * Each page table has 1024 entries × 8 bytes = 8KB (2 pages)
  */
 typedef struct {
-	unsigned long pte;
+	unsigned long long pte;
 } pte_t;
 
 typedef struct {
-	unsigned long pgd;
+	unsigned long long pgd;
 } pgd_t;
 
 typedef struct {
-	unsigned long pgprot;
+	unsigned long long pgprot;
 } pgprot_t;
 
 typedef struct page *pgtable_t;
